fixed ICMP latency and ping duration, organized Vue files
