+++ frontend/src/components/Dashboard/ServiceInfo.vue
@@ -17,7 +17,7 @@
                 <div v-if="false" class="row mb-4 align-content-center">
 
                     <div v-if="!service.online" class="col-3 text-left">
-                        <span class="text-danger font-5 font-weight-bold"></span>
+                        <span css="text-danger font-5 font-weight-bold"></span>
                         <span class="font-2 d-block">Current Downtime</span>
                     </div>
 
@@ -103,7 +103,7 @@
   import Checkin from '../../forms/Checkin';
   import FormIncident from '../../forms/Incident';
   import FormMessage from '../../forms/Message';
-  import ServiceFailures from './ServiceFailures';
+  import ServiceFailures from '../Service/ServiceFailures';
   import ServiceSparkLine from "./ServiceSparkLine";
   import Api from "../../API";
