+++ utils/utils_custom.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"os"
 	"os/exec"
+	"regexp"
 	"strconv"
 	"strings"
 	"syscall"
@@ -36,20 +37,27 @@ func DirWritable(path string) (bool, error) {
 	return true, nil
 }
 
-func Ping(address string, secondsTimeout int) error {
+func Ping(address string, secondsTimeout int) (int64, error) {
 	ping, err := exec.LookPath("ping")
 	if err != nil {
-		return err
+		return 0, err
 	}
 	out, _, err := Command(ping, address, "-c", "1", "-W", strconv.Itoa(secondsTimeout))
 	if err != nil {
-		return err
+		return 0, err
 	}
 	if strings.Contains(out, "Unknown host") {
-		return errors.New("unknown host")
+		return 0, errors.New("unknown host")
 	}
 	if strings.Contains(out, "100.0% packet loss") {
-		return errors.New("destination host unreachable")
+		return 0, errors.New("destination host unreachable")
 	}
-	return nil
+
+	r := regexp.MustCompile(`time=(.*) ms`)
+	strs := r.FindStringSubmatch(out)
+	if len(strs) < 2 {
+		return 0, errors.New("could not parse ping duration")
+	}
+	f, _ := strconv.ParseFloat(strs[1], 64)
+	return int64(f * 1000), nil
 }
